+++ src/main/java/insane96mcp/iguanatweaksreborn/modules/combat/feature/NoKnockbackFeature.java
@@ -1,12 +1,12 @@
 package insane96mcp.iguanatweaksreborn.modules.combat.feature;
 
 import com.google.common.collect.Multimap;
-import insane96mcp.iguanatweaksreborn.common.classutils.IdTagMatcher;
 import insane96mcp.iguanatweaksreborn.setup.Config;
 import insane96mcp.iguanatweaksreborn.setup.Strings;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
+import insane96mcp.insanelib.utils.IdTagMatcher;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.ai.attributes.Attribute;
 import net.minecraft.entity.ai.attributes.AttributeModifier;