+++ src/main/java/insane96mcp/iguanatweaksreborn/modules/farming/feature/NerfedBonemealFeature.java
@@ -1,6 +1,5 @@
 package insane96mcp.iguanatweaksreborn.modules.farming.feature;
 
-import insane96mcp.iguanatweaksreborn.common.classutils.IdTagMatcher;
 import insane96mcp.iguanatweaksreborn.modules.Modules;
 import insane96mcp.iguanatweaksreborn.modules.farming.FarmingModule;
 import insane96mcp.iguanatweaksreborn.modules.farming.classutils.CropsRequireWater;
@@ -9,6 +8,7 @@ import insane96mcp.iguanatweaksreborn.setup.Config;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
+import insane96mcp.insanelib.utils.IdTagMatcher;
 import insane96mcp.insanelib.utils.RandomHelper;
 import net.minecraft.block.BeetrootBlock;
 import net.minecraft.block.BlockState;